pyproject.toml
[project]
name = "telegrambotvk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pdfplumber>=0.11.8",
    "python-docx>=1.2.0",
    "python-telegram-bot>=22.5",
]
